Fix: Linter bugs
