+++ src/components/app/app.tsx
@@ -1,11 +1,14 @@
 import React from 'react';
-import MainPage, { loader as mainPageLoader } from '../../pages/main-page/main-page';
-import FavoritesPage, { loader as favoritesPageLoader } from '../../pages/favorites-page/favorites-page';
-import OfferPage, { loader as offerPageLoader } from '../../pages/offer-page/offer-page';
+import MainPage from '../../pages/main-page/main-page';
+import { loader as mainPageLoader } from '../../pages/main-page/main-page-loader';
+import FavoritesPage from '../../pages/favorites-page/favorites-page';
+import { loader as favoritesPageLoader } from '../../pages/favorites-page/favorites-page-loader';
+import OfferPage from '../../pages/offer-page/offer-page';
+import { loader as offerPageLoader } from '../../pages/offer-page/offer-page-loader';
 import LoginPage from '../../pages/login-page/login-page';
 import FavoritesEmptyPage from '../../pages/favorites-empty-page/favorites-empty-page';
 import Page404 from '../../pages/page-404/page-404';
-import { PrivateRoute, PublicRoute } from '../../pages/access-rout/access-rout';
+import { PrivateRoute, PublicRoute } from '../../pages/access-route/access-route';
 import { AppRoute, AuthorizationStatus } from '../../constants';
 import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import ScrollToTop from '../scroll-to-top/scroll-to-top';
