+++ src/constants/index.ts
@@ -1,5 +1,5 @@
 export {PROJECT_NAME, DEFAULT_CITY, CITIES, CitiesGPS, OFFER_TYPES, AuthorizationStatus} from './common';
-export {AppRoute} from './routs';
+export {AppRoute} from './routes';
 export { TILE_LAYER, TILE_COPYRIGHT } from './leaflet';
 export {TemporalData, LOCATION_RADIUS} from './mocks';
